Moves service helpers to separate files to promote SoC
