+++ src/utils/config/service-helpers.js
@@ -1,257 +1,11 @@
-import { promises as fs } from "fs";
-import path from "path";
+import { servicesFromDocker } from "./service-helpers/docker";
+import { servicesFromConfig } from "./service-helpers/config";
+import { servicesFromKubernetes } from "./service-helpers/kubernetes";
 
-import Docker from "dockerode";
-import yaml from "js-yaml";
-
-import checkAndCopyConfig, { CONF_DIR, getSettings, substituteEnvironmentVars } from "utils/config/config";
-import getDockerArguments from "utils/config/docker";
-import { getKubeConfig } from "utils/config/kubernetes";
-import * as shvl from "utils/config/shvl";
-import kubernetes from "utils/kubernetes/export";
 import createLogger from "utils/logger";
-
 const logger = createLogger("service-helpers");
 
-function parseServicesToGroups(services) {
-  if (!services) {
-    return [];
-  }
-
-  // map easy to write YAML objects into easy to consume JS arrays
-  return services.map((serviceGroup) => {
-    const name = Object.keys(serviceGroup)[0];
-    let groups = [];
-    const serviceGroupServices = [];
-    serviceGroup[name].forEach((entries) => {
-      const entryName = Object.keys(entries)[0];
-      if (!entries[entryName]) {
-        logger.warn(`Error parsing service "${entryName}" from config. Ensure required fields are present.`);
-        return;
-      }
-      if (Array.isArray(entries[entryName])) {
-        groups = groups.concat(parseServicesToGroups([{ [entryName]: entries[entryName] }]));
-      } else {
-        serviceGroupServices.push({
-          name: entryName,
-          ...entries[entryName],
-          weight: entries[entryName].weight || serviceGroupServices.length * 100, // default weight
-          type: "service",
-        });
-      }
-    });
-    return {
-      name,
-      type: "group",
-      services: serviceGroupServices,
-      groups,
-    };
-  });
-}
-
-export async function servicesFromConfig() {
-  checkAndCopyConfig("services.yaml");
-
-  const servicesYaml = path.join(CONF_DIR, "services.yaml");
-  const rawFileContents = await fs.readFile(servicesYaml, "utf8");
-  const fileContents = substituteEnvironmentVars(rawFileContents);
-  const services = yaml.load(fileContents);
-  return parseServicesToGroups(services);
-}
-
-export async function servicesFromDocker() {
-  checkAndCopyConfig("docker.yaml");
-
-  const dockerYaml = path.join(CONF_DIR, "docker.yaml");
-  const rawDockerFileContents = await fs.promises.readFile(dockerYaml, "utf8");
-  const dockerFileContents = substituteEnvironmentVars(rawDockerFileContents);
-  const servers = yaml.load(dockerFileContents);
-
-  if (!servers) {
-    return [];
-  }
-
-  const { instanceName } = getSettings();
-
-  const serviceServers = await Promise.all(
-    Object.keys(servers).map(async (serverName) => {
-      try {
-        const isSwarm = !!servers[serverName].swarm;
-        const docker = new Docker(getDockerArguments(serverName).conn);
-        const listProperties = { all: true };
-        const containers = await (isSwarm
-          ? docker.listServices(listProperties)
-          : docker.listContainers(listProperties));
-
-        // bad docker connections can result in a <Buffer ...> object?
-        // in any case, this ensures the result is the expected array
-        if (!Array.isArray(containers)) {
-          return [];
-        }
-
-        const discovered = [];
-        containers.forEach((container) => {
-          const generator = createServicesFromContainer(container);
-          let result = generator.next();
-          while (!result.done) {
-            if (result.value !== null) discovered.push(result.value);
-            result = generator.next();
-          }
-        });
-
-        function* createServicesFromContainer(container) {
-          let constructedService = null;
-          const containerLabels = isSwarm ? shvl.get(container, "Spec.Labels") : container.Labels;
-          const containerName = isSwarm ? shvl.get(container, "Spec.Name") : container.Names[0];
-
-          for (const [labelName, labelValue] of Object.entries(containerLabels)) {
-            if (!labelName.startsWith("homepage.")) {
-              continue;
-            }
-
-            let propertyPath = labelName.replace("homepage.", "");
-            if (instanceName && propertyPath.startsWith(`instance.${instanceName}.`)) {
-              propertyPath = propertyPath.replace(`instance.${instanceName}.`, "");
-            } else if (propertyPath.startsWith("instance.")) {
-              continue;
-            }
-
-            if (!constructedService) {
-              constructedService = {
-                container: containerName.replace(/^\//, ""),
-                server: serverName,
-                type: "service",
-              };
-            }
-
-            let substitutedVal = substituteEnvironmentVars(labelValue);
-            if (propertyPath === "widget.version") {
-              substitutedVal = parseInt(substitutedVal, 10);
-            }
-
-            shvl.set(constructedService, propertyPath, substitutedVal);
-          }
-
-          const tryCreateService = service => {
-            if (service && (!service.name || !service.group) && !service.services) {
-              logger.error(
-                `Error constructing service using homepage labels for container '${containerName.replace(
-                  /^\//,
-                  "",
-                )}'. Ensure required labels are present.`,
-              );
-              logger.error(service);
-              return null;
-            }
-
-            return service;
-          }
-
-          if (constructedService && constructedService.services) {
-            for (const service of constructedService.services) {
-              const resultingService = structuredClone(constructedService);
-              delete resultingService.services;
-              yield tryCreateService(Object.assign(resultingService, service));
-            }
-            return;
-          }
-
-          yield tryCreateService(constructedService);
-        }
-
-        return { server: serverName, services: discovered.filter((filteredService) => filteredService) };
-      } catch (e) {
-        logger.error("Error getting services from Docker server '%s': %s", serverName, e);
-
-        // a server failed, but others may succeed
-        return { server: serverName, services: [] };
-      }
-    }),
-  );
-
-  const mappedServiceGroups = [];
-
-  serviceServers.forEach((server) => {
-    server.services.forEach((serverService) => {
-      let serverGroup = mappedServiceGroups.find((searchedGroup) => searchedGroup.name === serverService.group);
-      if (!serverGroup) {
-        mappedServiceGroups.push({
-          name: serverService.group,
-          services: [],
-        });
-        serverGroup = mappedServiceGroups[mappedServiceGroups.length - 1];
-      }
-
-      const { name: serviceName, group: serverServiceGroup, ...pushedService } = serverService;
-      const result = {
-        name: serviceName,
-        ...pushedService,
-      };
-
-      serverGroup.services.push(result);
-    });
-  });
-
-  return mappedServiceGroups;
-}
-
-export async function servicesFromKubernetes() {
-  const { instanceName } = getSettings();
-
-  checkAndCopyConfig("kubernetes.yaml");
-
-  try {
-    const kc = getKubeConfig();
-    if (!kc) {
-      return [];
-    }
-
-    // resource lists
-    const [ingressList, traefikIngressList, httpRouteList] = await Promise.all([
-      kubernetes.listIngress(),
-      kubernetes.listTraefikIngress(),
-      kubernetes.listHttpRoute(),
-    ]);
-
-    const resources = [...ingressList, ...traefikIngressList, ...httpRouteList];
-
-    if (!resources) {
-      return [];
-    }
-    const services = await Promise.all(
-      resources
-        .filter((resource) => kubernetes.isDiscoverable(resource, instanceName))
-        .map(async (resource) => kubernetes.constructedServiceFromResource(resource)),
-    );
-
-    // map service groups
-    const mappedServiceGroups = services.reduce((groups, serverService) => {
-      let serverGroup = groups.find((group) => group.name === serverService.group);
-
-      if (!serverGroup) {
-        serverGroup = {
-          name: serverService.group,
-          services: [],
-        };
-        groups.push(serverGroup);
-      }
-
-      const { name: serviceName, group: _, ...pushedService } = serverService;
-
-      serverGroup.services.push({
-        name: serviceName,
-        ...pushedService,
-      });
-
-      return groups;
-    }, []);
-
-    return mappedServiceGroups;
-  } catch (e) {
-    if (e) logger.error(e);
-    throw e;
-  }
-}
+export {servicesFromDocker, servicesFromConfig, servicesFromKubernetes};
 
 export function cleanServiceGroups(groups) {
   return groups.map((serviceGroup) => ({
